+++ test/unit/fuzz-tools.test.js
@@ -519,6 +519,7 @@ import * as rachs1V541 from '../../lib/rachs1-v541.js';
 import * as twstrsSeverityV542 from '../../lib/twstrs-severity-v542.js';
 import * as saveScoreV543 from '../../lib/save-score-v543.js';
 import * as nemsV544 from '../../lib/nems-v544.js';
+import * as palmCoeinV545 from '../../lib/palm-coein-v545.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -983,6 +984,7 @@ const MODULES = {
   'twstrs-severity-v542.js': twstrsSeverityV542,
   'save-score-v543.js': saveScoreV543,
   'nems-v544.js': nemsV544,
+  'palm-coein-v545.js': palmCoeinV545,
   'tb-testing.js': tbTesting,
 };